+++ lib/glimmer-dsl-web/samples/hello/hello_input_date_time_data_binding.rb
@@ -32,76 +32,47 @@ unless Object.const_defined?(:TimePresenter) # this is only needed in sample sel
   end
 end
 
-unless Object.const_defined?(:HelloInputDateTime) # this is only needed in sample selector app due to file reloading, but not in real apps.
-  class HelloInputDateTime
+unless Object.const_defined?(:HelloInputDateTimeDataBinding) # this is only needed in sample selector app due to file reloading, but not in real apps.
+  class HelloInputDateTimeDataBinding
     include Glimmer::Web::Component
     
     before_render do
       @time_presenter = TimePresenter.new
     end
     
-    after_render do
-      observe(@time_presenter, :date_time) do |new_date_time|
-        @date_time_field.value = new_date_time
-        @date_field.value = new_date_time
-        @time_field.value = new_date_time
-        @month_field.value = new_date_time
-        @week_field.value = @time_presenter.week_string
-        @time_string_field.value = @time_presenter.date_time_string
-      end
-    end
-    
     markup {
       div {
         div(style: 'display: grid; grid-auto-columns: 130px 260px;') { |container_div|
           label('Date Time: ', for: 'date-time-field')
-          @date_time_field = input(id: 'date-time-field', type: 'datetime-local') {
-            value @time_presenter.date_time
-            
-            oninput do |event|
-              @time_presenter.date_time = @date_time_field.value
-            end
+          input(id: 'date-time-field', type: 'datetime-local') {
+            # Bidirectional Data-Binding with <=> ensures input.value and @time_presenter.date_time
+            # automatically stay in sync when either side changes
+            value <=> [@time_presenter, :date_time]
           }
           
           label('Date: ', for: 'date-field')
-          @date_field = input(id: 'date-field', type: 'date') {
-            value @time_presenter.date_time
-            
-            oninput do |event|
-              @time_presenter.date_time = @date_field.value
-            end
+          input(id: 'date-field', type: 'date') {
+            value <=> [@time_presenter, :date_time]
           }
     
           label('Time: ', for: 'time-field')
-          @time_field = input(id: 'time-field', type: 'time') {
-            value @time_presenter.date_time
-            
-            oninput do |event|
-              @time_presenter.date_time = @time_field.value
-            end
+          input(id: 'time-field', type: 'time') {
+            value <=> [@time_presenter, :date_time]
           }
-  
+    
           label('Month: ', for: 'month-field')
-          @month_field = input(id: 'month-field', type: 'month') {
-            value @time_presenter.date_time
-            
-            oninput do |event|
-              @time_presenter.date_time = @month_field.value
-            end
+          input(id: 'month-field', type: 'month') {
+            value <=> [@time_presenter, :date_time]
           }
-  
+    
           label('Week: ', for: 'week-field')
-          @week_field = input(id: 'week-field', type: 'week', disabled: true) {
-            value @time_presenter.week_string
+          input(id: 'week-field', type: 'week', disabled: true) {
+            value <= [@time_presenter, :week_string, computed_by: :date_time]
           }
-  
+    
           label('Time String: ', for: 'time-string-field')
-          @time_string_field = input(id: 'time-string-field', type: 'text') {
-            value @time_presenter.date_time_string
-          
-            oninput do |event|
-              @time_presenter.date_time_string = @time_string_field.value
-            end
+          input(id: 'time-string-field', type: 'text') {
+            value <=> [@time_presenter, :date_time_string, computed_by: :date_time]
           }
           
           style {
@@ -122,5 +93,5 @@ unless Object.const_defined?(:HelloInputDateTime) # this is only needed in sampl
 end
 
 Document.ready? do
-  HelloInputDateTime.render
+  HelloInputDateTimeDataBinding.render
 end